Streamline minimizing
Fedora images already minimize themselves during creation, so this is
(mostly) duplicate work.

+++ scripts/minimize.sh
@@ -6,12 +6,12 @@ set -x
 rm -rf /home/vagrant/.ansible
 sudo rm -rf /root/.ansible
 
-#if [[ "${PACKER_BUILD_NAME}" != *"silverblue"* ]]; then
+if [[ ! -e /etc/fedora-release ]]; then
 	sudo dd if=/dev/zero of=/var/EMPTY bs=1M || :
 	sync
 	sudo rm /var/EMPTY
 	sync
-#fi
+fi
 
 # In CentOS 7, blkid returns duplicate devices
 swap_device_uuid=`sudo /sbin/blkid -t TYPE=swap -o value -s UUID | uniq`
